requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
Django==3.1.1
django-crontab==0.7.1
django-rest-knox==4.1.0
djangorestframework==3.11.1
idna==2.10
pkg-resources==0.0.0
pycparser==2.20
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
